--- moiza-domain/src/main/java/com/moiza/moizaspringbootserver/auth/EmailCode.java
@@ -1,16 +1,0 @@
-package com.moiza.moizaspringbootserver.auth;
-
-import com.moiza.moizaspringbootserver.annotation.Aggregate;
-import com.moiza.moizaspringbootserver.auth.type.Type;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@Aggregate
-public class EmailCode {
-
-    private final String email;
-    private final String authCode;
-    private final Type type;
-}
