+++ moiza-domain/src/main/java/com/moiza/moizaspringbootserver/auth/domain/RefreshToken.java
@@ -1,4 +1,4 @@
-package com.moiza.moizaspringbootserver.auth;
+package com.moiza.moizaspringbootserver.auth.domain;
 
 import com.moiza.moizaspringbootserver.annotation.Aggregate;
 import lombok.Builder;
