+++ moiza-domain/src/main/java/com/moiza/moizaspringbootserver/auth/domain/type/Type.java
@@ -1,0 +1,7 @@
+package com.moiza.moizaspringbootserver.auth.domain.type;
+
+public enum Type {
+    SIGNUP,
+    ID,
+    PASSWORD
+}
