--- moiza-domain/src/main/java/com/moiza/moizaspringbootserver/auth/type/Type.java
@@ -1,7 +1,0 @@
-package com.moiza.moizaspringbootserver.auth.type;
-
-public enum Type {
-    SIGNUP,
-    ID,
-    PASSWORD
-}
